migrating env search logic
 - migrating from Commands.UserToolchain to Utility.Process
 - migrated tests to match logic move
 - updated XCTests to reflect migrated test cases
