+++ Tests/CommandsTests/XCTestManifests.swift
@@ -16,7 +16,6 @@ public func allTests() -> [XCTestCaseEntry] {
         testCase(BuildToolTests.allTests),
         testCase(PackageToolTests.allTests),
         testCase(TestToolTests.allTests),
-        testCase(UserToolchainTests.allTests),
     ]
 }
 #endif
