+++ Tests/UtilityTests/miscTests.swift
@@ -10,6 +10,7 @@
 
 import XCTest
 import TestSupport
+import Basic
 @testable import Utility
 
 class miscTests: XCTestCase {
@@ -41,8 +42,56 @@ class miscTests: XCTestCase {
         }
     }
 
+    func testExecutableLookup() throws {
+        mktmpdir { path in
+            
+            let pathEnv1 = path.appending(component: "pathEnv1")
+            try localFileSystem.createDirectory(pathEnv1)
+            let pathEnvClang = pathEnv1.appending(component: "clang")
+            try localFileSystem.writeFileContents(pathEnvClang, bytes: "")
+            let pathEnv = [path.appending(component: "pathEnv2"), pathEnv1]
+            
+            // nil and empty string should fail.
+            XCTAssertNil(lookupExecutablePath(inEnvValue: nil, currentWorkingDirectory: path, searchPaths: pathEnv))
+            XCTAssertNil(lookupExecutablePath(inEnvValue: "", currentWorkingDirectory: path, searchPaths: pathEnv))
+            
+            // Absolute path to a binary should return it.
+            var exec = lookupExecutablePath(inEnvValue: pathEnvClang.asString, currentWorkingDirectory: path, searchPaths: pathEnv)
+            XCTAssertEqual(exec, pathEnvClang)
+            
+            // This should lookup from PATH variable since executable is not present in cwd.
+            exec = lookupExecutablePath(inEnvValue: "clang", currentWorkingDirectory: path, searchPaths: pathEnv)
+            XCTAssertEqual(exec, pathEnvClang)
+            
+            // Create the binary relative to cwd.
+            let clang = path.appending(component: "clang")
+            try localFileSystem.writeFileContents(clang, bytes: "")
+            // We should now find clang which is in cwd.
+            exec = lookupExecutablePath(inEnvValue: "clang", currentWorkingDirectory: path, searchPaths: pathEnv)
+            XCTAssertEqual(exec, clang)
+        }
+    }
+    
+    func testEnvSearchPaths() throws {
+        let cwd = AbsolutePath("/dummy")
+        let paths = getEnvSearchPaths(pathString: "something:.:abc/../.build/debug:/usr/bin:/bin/", currentWorkingDirectory: cwd)
+        XCTAssertEqual(paths, ["/dummy/something", "/dummy", "/dummy/.build/debug", "/usr/bin", "/bin"].map(AbsolutePath.init))
+    }
+    
+    func testEmptyEnvSearchPaths() throws {
+        let cwd = AbsolutePath("/dummy")
+        let paths = getEnvSearchPaths(pathString: "", currentWorkingDirectory: cwd)
+        XCTAssertEqual(paths, [])
+        
+        let nilPaths = getEnvSearchPaths(pathString: nil, currentWorkingDirectory: cwd)
+        XCTAssertEqual(nilPaths, [])
+    }
+
     static var allTests = [
         ("testClangVersionOutput", testClangVersionOutput),
         ("testVersion", testVersion),
+        ("testExecutableLookup", testExecutableLookup),
+        ("testEnvSearchPaths", testEnvSearchPaths),
+        ("testEmptyEnvSearchPaths", testEnvSearchPaths),
     ]
 }
